Extracted request handler from Custom404Handler.
