+++ src/Core/Configuration/FileNotFoundMode.cs
@@ -1,0 +1,18 @@
+namespace BVNetwork.NotFound.Core.Configuration
+{
+    public enum FileNotFoundMode
+    {
+        /// <summary>
+        ///
+        /// </summary>
+        On,
+        /// <summary>
+        ///
+        /// </summary>
+        Off,
+        /// <summary>
+        ///
+        /// </summary>
+        RemoteOnly
+    };
+}
