+++ src/Core/Configuration/LoggerMode.cs
@@ -1,0 +1,7 @@
+namespace BVNetwork.NotFound.Core.Configuration
+{
+    public enum LoggerMode
+    {
+        On, Off
+    };
+}
